+++ libs/app-serialport/devices/src/lib/harmony-device/serial-port-harmony-msc-device.ts
@@ -8,8 +8,8 @@ import {
   SerialPortDeviceOptions,
 } from "../serial-port-device"
 import { SerialPortDeviceType, SerialPortRequest } from "app-serialport/models"
-import { CommonDeviceResponseParser } from "../common/common-device-response-parser"
-import { commonDeviceRequestParser } from "../common/common-device-request-parser"
+import { CommonDeviceResponseParser } from "../shared/common-device-response-parser"
+import { commonDeviceRequestParser } from "../shared/common-device-request-parser"
 
 export class SerialPortHarmonyMscDevice extends SerialPortDevice {
   static matchingVendorIds = ["3310"]
