+++ libs/app-serialport/devices/src/lib/shared/common-device-request-parser.test.ts
