+++ libs/app-serialport/devices/src/lib/shared/common-device-response-parser.ts
